pyproject.toml
[project]
name = "czii-model-contribution"
version = "0.1.0"
description = "7th place winning model of the czii object identification competition"
requires-python = ">=3.10"
dependencies = [
    "cryoet-data-portal>=4.4.1",
    "matplotlib>=3.10.3",
    "ndjson>=0.3.1",
    "numpy>=2.2.5",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scipy>=1.15.3",
    "segmentation-models-pytorch>=0.5.0",
    "timm>=1.0.15",
    "timm-3d>=1.0.1",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "zarr>=2.18.3",
]
